# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTaskRequest:
    properties:
      description:
        description: Description is an optional detailed description of the task.
        example: Create a RESTful endpoint for user authentication.
        maxLength: 65535
        type: string
      due_date:
        description: DueDate is the optional due date of the task.
        example: "2025-04-20T00:00:00Z"
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        description: Priority is the priority level of the task.
        enum:
        - HIGH
        - MEDIUM
        - LOW
        - CRITICAL
        example: HIGH
      sprint_id:
        description: SprintID is the ID of the sprint this task belongs to.
        example: 1
        minimum: 1
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        description: Status is the current status of the task.
        enum:
        - TO_DO
        - IN_PROGRESS
        - REVIEW
        - DONE
        - BLOCKED
        example: TO_DO
      title:
        description: Title is the title of the task.
        example: Implement login API
        maxLength: 255
        minLength: 2
        type: string
    required:
    - sprint_id
    - title
    type: object
...
  dto.ErrorResponse:
    properties:
      details: {}
      message:
        example: An error occurred
        type: string
    type: object
  dto.GenericSuccessResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.TaskResponse:
    properties:
      assignee_first_name:
        description: AssigneeFirstName is the optional first name of the assignee.
        example: John
        type: string
      assignee_id:
        description: AssigneeID is the optional ID of the user assigned to the task.
        example: 42
        type: integer
      assignee_last_name:
        description: AssigneeLastName is the optional last name of the assignee.
        example: Doe
        type: string
      description:
        description: Description is the detailed description of the task.
        example: Create a RESTful endpoint for user authentication.
        type: string
      due_date:
        description: DueDate is the optional due date of the task.
        example: "2025-04-20T00:00:00Z"
        type: string
      id:
        description: ID is the unique identifier of the task.
        example: 101
        type: integer
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        description: Priority is the priority level of the task.
        example: HIGH
      project_id:
        description: ProjectID is the ID of the project this task belongs to.
        example: 1
        type: integer
      project_name:
        description: ProjectName is the name of the project this task belongs to.
        example: Website Redesign
        type: string
      sprint_id:
        description: SprintID is the ID of the sprint this task belongs to.
        example: 1
        type: integer
      sprint_name:
        description: SprintName is the name of the sprint this task belongs to.
        example: Sprint 1
        type: string
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        description: Status is the current status of the task.
        example: IN_PROGRESS
      title:
        description: Title is the title of the task.
        example: Implement login API
        type: string
    type: object
  dto.TaskSliceSuccessResponse:
    properties:
      count:
        example: 5
        type: integer
      data:
        items:
          $ref: '#/definitions/dto.TaskResponse'
        type: array
      message:
        example: Items found successfully
        type: string
    type: object
  dto.TaskSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/dto.TaskResponse'
      message:
        example: Operation successful
        type: string
    type: object
...
  dto.UpdateTaskRequest:
    properties:
      description:
        description: Description is the optional new description of the task.
        example: Modify endpoint to include JWT.
        maxLength: 65535
        type: string
      due_date:
        description: DueDate is the optional new due date of the task.
        example: "2025-04-25T00:00:00Z"
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/models.TaskPriority'
        description: Priority is the optional new priority level of the task.
        enum:
        - HIGH
        - MEDIUM
        - LOW
        - CRITICAL
        example: MEDIUM
      status:
        allOf:
        - $ref: '#/definitions/models.TaskStatus'
        description: Status is the optional new status of the task.
        enum:
        - TO_DO
        - IN_PROGRESS
        - REVIEW
        - DONE
        - BLOCKED
        example: REVIEW
      title:
        description: Title is the optional new title of the task.
        example: Update login API
        maxLength: 255
        minLength: 2
        type: string
    type: object
...
  models.TaskPriority:
    enum:
    - HIGH
    - MEDIUM
    - LOW
    - CRITICAL
    type: string
    x-enum-varnames:
    - HighPriority
    - MediumPriority
    - LowPriority
    - CriticalPriority
  models.TaskStatus:
    enum:
    - TO_DO
    - IN_PROGRESS
    - REVIEW
    - DONE
    - BLOCKED
    type: string
    x-enum-varnames:
    - ToDoTask
    - InProgressTask
    - ReviewTask
    - DoneTask
    - BlockedTask
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample swagger for Fiber
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
...
  /projects/{projectId}/tasks:
    get:
      description: Retrieves all tasks associated with a specific project
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Tasks found
          schema:
            $ref: '#/definitions/dto.TaskSliceSuccessResponse'
        "400":
          description: Bad request - Invalid project ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Project not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get tasks by project ID
      tags:
      - Tasks
...
  /tasks:
    get:
      description: Retrieves tasks based on optional query parameters (id, title,
        status, priority, due_date_before)
      parameters:
      - description: Task ID
        in: query
        name: id
        type: integer
      - description: Task title
        in: query
        name: title
        type: string
      - description: Task status
        enum:
        - OPEN
        - IN_PROGRESS
        - DONE
        in: query
        name: status
        type: string
      - description: Task priority
        enum:
        - LOW
        - MEDIUM
        - HIGH
        in: query
        name: priority
        type: string
      - description: 'Due date before (format: YYYY-MM-DD)'
        in: query
        name: due_date_before
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Tasks found
          schema:
            $ref: '#/definitions/dto.TaskSliceSuccessResponse'
        "400":
          description: Bad request - Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Find tasks with filters
      tags:
      - Tasks
    post:
      consumes:
      - application/json
      description: Creates a new task with the provided details for a specific sprint
      parameters:
      - description: Task creation request
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Task created successfully
          schema:
            $ref: '#/definitions/dto.TaskSuccessResponse'
        "400":
          description: Bad request - Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Sprint not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
  /tasks/{taskId}:
    delete:
      description: Deletes a specific task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Task deleted successfully
          schema:
            $ref: '#/definitions/dto.GenericSuccessResponse'
        "400":
          description: Bad request - Invalid task ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Task not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - Tasks
    get:
      description: Retrieves details of a specific task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Task found
          schema:
            $ref: '#/definitions/dto.TaskSuccessResponse'
        "400":
          description: Bad request - Invalid task ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Task not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Updates the details of an existing task
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: Task update request
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Task updated
          schema:
            $ref: '#/definitions/dto.TaskSuccessResponse'
        "400":
          description: Bad request - Invalid input or task ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Task not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - Tasks
  /tasks/{taskId}/user/{userId}:
    post:
      description: Assigns a specific task to a user
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: integer
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Task assigned successfully
          schema:
            $ref: '#/definitions/dto.GenericSuccessResponse'
        "400":
          description: Bad request - Invalid IDs or user not in project
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - User not authorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - Task or user not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Assign task to user
      tags:
      - Tasks
...
  /users/{userId}/tasks:
    get:
      description: Retrieves all tasks assigned to a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Tasks found
          schema:
            $ref: '#/definitions/dto.TaskSliceSuccessResponse'
        "400":
          description: Bad request - Invalid user ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not found - User not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get tasks by user ID
      tags:
      - Tasks
schemes:
- http
- https
swagger: "2.0"